+++ src/Method/EditForumTopic.php
@@ -1,0 +1,52 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Vjik\TelegramBot\Api\Method;
+
+use Vjik\TelegramBot\Api\ParseResult\ValueHelper;
+use Vjik\TelegramBot\Api\Request\HttpMethod;
+use Vjik\TelegramBot\Api\Request\TelegramRequestWithResultPreparingInterface;
+
+/**
+ * @see https://core.telegram.org/bots/api#editforumtopic
+ */
+final readonly class EditForumTopic implements TelegramRequestWithResultPreparingInterface
+{
+    public function __construct(
+        private int|string $chatId,
+        private int $messageThreadId,
+        private ?string $name = null,
+        private ?string $iconCustomEmojiId = null,
+    ) {
+    }
+
+    public function getHttpMethod(): HttpMethod
+    {
+        return HttpMethod::POST;
+    }
+
+    public function getApiMethod(): string
+    {
+        return 'editForumTopic';
+    }
+
+    public function getData(): array
+    {
+        return array_filter(
+            [
+                'chat_id' => $this->chatId,
+                'message_thread_id' => $this->messageThreadId,
+                'name' => $this->name,
+                'icon_custom_emoji_id' => $this->iconCustomEmojiId,
+            ],
+            static fn(mixed $value): bool => $value !== null,
+        );
+    }
+
+    public function prepareResult(mixed $result): true
+    {
+        ValueHelper::assertTrueResult($result);
+        return $result;
+    }
+}